requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==2.4.1
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.1.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
hydra-core==1.3.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kiwisolver==1.4.7
lightgbm==4.5.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
missingno==0.5.2
nest-asyncio==1.6.0
numpy==2.1.3
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.19.1
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
text-unidecode==1.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.0
wcwidth==0.2.13
webencodings==0.5.1
